More work on tests
