+++ test/DynamoCoreTests/DynamoCoreTests.csproj
@@ -87,6 +87,7 @@
     <Compile Include="DummyNodeTests.cs" />
     <Compile Include="ImperativeClodeBlockTest.cs" />
     <Compile Include="InheritanceTests.cs" />
+    <Compile Include="InputOutputDataTests.cs" />
     <Compile Include="libGPreloaderTests.cs" />
     <Compile Include="ListAtLevelTests.cs" />
     <Compile Include="MultiOutputNodeTest.cs" />
